pyproject.toml
[project]
name = "neurips"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.72.0",
    "nltk>=3.9.2",
    "pandas>=2.3.3",
    "plotly>=6.4.0",
    "sentence-transformers>=5.1.2",
    "umap-learn>=0.5.9.post2",
    "wordcloud>=1.9.4",
]
